edit trains table numero_carrozze column

+++ database/migrations/2024_05_10_132711_create_trains_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->time('orario_partenza', $precision = 0);
             $table->time('orario_arrivo', $precision = 0);
             $table->string('codice_treno', 15);
-            $table->tinyInteger('numero_carrozze');
+            $table->smallInteger('numero_carrozze');
             $table->boolean('in_orario')->default(1)->nullable();
             $table->boolean('cancellato')->default(1)->nullable();
             $table->timestamps();
